% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{do_meta_analysis}
\alias{do_meta_analysis}
\title{Do metaanalysis}
\usage{
do_meta_analysis(
  plot_height = 5,
  plot_width = 7,
  plot_ukb = TRUE,
  plot_aou = TRUE,
  plot_biovu = TRUE
)
}
\value{
none, just side effects of tables and plots
}
\description{
Do metaanalysis
}
\details{
functions to carry out meta analysis combining three biobanks
}
