% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{one_cohort_results_plot}
\alias{one_cohort_results_plot}
\title{one cohort results plot}
\usage{
one_cohort_results_plot(
  results,
  title = "Hazard Ratio of CHIP for Incident RA",
  andor = "and",
  p_or_q = "p",
  do_exp = TRUE,
  plot_top = 10
)
}
\arguments{
\item{results}{the results to plot}

\item{title}{the title of the plot}

\item{andor}{whether to plot and-based definition or or-based definition of disease}

\item{p_or_q}{plot using p values or q values}
}
\value{
the plot
}
\description{
one cohort results plot
}
