% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{one_cohort_specific_results_plot}
\alias{one_cohort_specific_results_plot}
\title{one cohort specific results plot}
\usage{
one_cohort_specific_results_plot(results, title, p_or_q = "p")
}
\arguments{
\item{results}{the results to plot}

\item{title}{the title of the plot}

\item{p_or_q}{plot using p values or q values}
}
\value{
the plot
}
\description{
one cohort specific results plot
}
