% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_surv.R
\name{prepare_baseline_data}
\alias{prepare_baseline_data}
\title{prepare baseline data}
\usage{
prepare_baseline_data(demographics, chip_calls)
}
\arguments{
\item{demographics}{tibble with baseline demographic data. See details.}

\item{chip_calls}{tibble with chip call status. See details}
}
\value{
baseline data prepped for survival analysis
}
\description{
prepare baseline data
}
\details{
demographics includes: person_id (character), sex (factor),
race (factor), ever_smoker (logical), biosample_date (date),
age_at_biosample (double), age2 (double), date_first_heme_ca
(date or NA), and date_last_dx (date or NA).

chip_calls includes: person_id (character), chip_gene (string or NA), and
AF (double or NA).
}
