% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_surv.R
\name{prepare_ra_outcomes}
\alias{prepare_ra_outcomes}
\title{Prepare RA Outcomes}
\usage{
prepare_ra_outcomes(
  dx_code_counts_df,
  serostatus_df,
  ra_meds_df,
  min_num_dx_codes
)
}
\arguments{
\item{dx_code_counts_df}{tibble of diagnosis codes. See details.}

\item{serostatus_df}{tibble of sero-status. See details.}

\item{ra_meds_df}{tibble of medications. See details.}

\item{min_num_dx_codes}{minimum number of occurrences of a specific
diagnosis code for a patient to be considered to have that disease}
}
\value{
tibble with ten columns: person_id and nine RA phenotype
definitions, three for RA, three for SPRA, and three for SNRA. The three are
(1) sensitive, based on diagnosis codes only, (2) moderate,based on
diagnosis code and either labs or meds, and (3) specific, based on diagnosis
codes, labs, and meds
}
\description{
Prepare RA Outcomes
}
\details{
dx_code_counts_df is a tibble with the four columns: person_id, dx,
dx_count, date_first_dx. The values in dx can be 'RA', 'SPRA', or
'SNRA'. dx_count is an integer indicating how many times that diagnosis
code was seen on that patient. date_first_dx is the date that diagnosis was
first seen on that patient.

serostatus_df is a tibble with two columns: person_id, and seropos.
The values in seropos can be TRUE and FALSE.

ra_meds_df is a tibble with two columns: person_id and treated_w_ra_meds.
The values in treated_w_ra_meds can be TRUE and FALSE.
}
