% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_surv.R
\name{prepare_baseline_data}
\alias{prepare_baseline_data}
\title{prepare baseline data}
\usage{
prepare_baseline_data(
  demographics,
  chip_calls,
  min_num_dx = 5,
  thresholds = c(`05` = 0.05, `10` = 0.1, `15` = 0.15, `20` = 0.2),
  genes = c("DNMT3A", "TET2")
)
}
\arguments{
\item{demographics}{tibble with baseline demographic data. See details.}

\item{chip_calls}{tibble with chip call status. See details}

\item{min_num_dx}{minimum number of diagnoses a patient must have to be studied}
}
\value{
baseline data prepped for survival analysis
}
\description{
prepare baseline data
}
\details{
demographics includes: person_id (character), sex (factor),
race (factor), ever_smoker (logical), biosample_date (date),
age_at_biosample (double), age2 (double), date_first_heme_ca
(date or NA), and date_last_dx (date or NA).

chip_calls includes: person_id (character), chip_gene (string or NA), and
AF (double or NA).
}
